backlight: lm3630: Expose min/max brightness levels to sysfs
This could be helpful for people who don't like the original limits
and have a desire to change them without recompiling the kernel or
manipulating with DT.

Change-Id: Iaccfff66d1c2d4b1ab1b0c2ebaf5276a1c6a9687
Signed-off-by: Alex Saiko <solcmdr@gmail.com>

backlight: lm3630: Fix ->bl_set_intensity() call and adjust limits

A generic set_intensity function had been broken since 850aa44bf14e
("backlight: lm3630: clean and refactor code"), when an erroneous check
of the new brightness with itself was introduced. Backlight framework
updates the bd->props.brightness value right before the bl_set_intensity()
call, hence this value cannot change during the setting. Because of this,
a check of an equality between the required level (bd->props.brightness)
and a current level (bd->props.brightness too) almost always fails, as
they are obviously the same most of the time.

To fix this, use a local current brightness indicator just like it was
before. By default, that indicator is equal to default brightness level.

This local brightness level is also used in backlight limits sysfs
functions to setup the backlight panel after the bound change. This way we
achieve 2 goals in one step:
- An update of a backlight after the bound change happens immediately, so
  there is no requirement to pdown/pon the panel anymore.
- Remote bl_dev->props.brightness value also updates instantly as it is
  synchronized with cur_brightness value in lm3630_set_main_current_level()
  call, which is used in these sysfs nodes now.

Change-Id: If59c1c3da6e07b3025932a3762e1e58f419824c6
Signed-off-by: Alex Saiko <solcmdr@gmail.com>